+++ src/pages/Landing/lib/landingData.ts
