+++ src/pages/Landing/ui/CtaBanner.tsx
@@ -1,7 +1,7 @@
 // Figma: https://www.figma.com/design/4rJmEFUU2HMWVy3qUcYZRs/%EC%A0%9C%EB%AA%A9-%EC%97%86%EC%9D%8C?node-id=1-12014
 // Figma-states: landing
 
-import type { LandingTabData } from "./landingData";
+import type { LandingTabData } from "../lib/landingData";
 
 interface CtaBannerProps {
   tabData: LandingTabData;
