+++ src/pages/Landing/ui/TabSelector.tsx
@@ -1,8 +1,8 @@
 // Figma: https://www.figma.com/design/4rJmEFUU2HMWVy3qUcYZRs/%EC%A0%9C%EB%AA%A9-%EC%97%86%EC%9D%8C?node-id=1-12014
 // Figma-states: landing
 
-import type { LandingTab } from './landingData';
-import { LANDING_TABS } from './landingData';
+import type { LandingTab } from '../lib/landingData';
+import { LANDING_TABS } from '../lib/landingData';
 
 interface TabSelectorProps {
   activeTab: LandingTab;
